dataimporter-src/requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.29
pandas==2.2.1
psycopg2-binary==2.9.9
ipython
ipykernel